Update timestamp validation in ValidationService
This commit updates the timestamp validation in the _ValidationSerivce_
to exclude milliseconds as part of the check on whether the application
and rep order timestamps are equal. The _ValidationService_ is also
slightly refactored in this process to make it clearer that if the
application timestamp is null then there isn't anything further to
validate with the timestamps.

This commit also adds some new test cases which specifically cover the
timestamp validation logic, as this was previously only covered for
failing cases, with the passing cases covered as part of the wider tests
on the _validate_ method.