Create FeatureDecisionService to determine feature branching logic
This commit introduces a new `FeatureDecisionService` which uses the
recently added feature flags in the `UserSummaryDTO` to decide whether
a specific feature is enabled for any given request. Currently, this
service contains a single method for determining if the CCC Service
is enabled.

The `isC3Enabled` flag in the `WorkflowRequest` class will be removed
in a later commit once all live code is using the new service added
as part of this commit.