Add check for MaatPostAssessmentProcessing flag to FeatureDecisionService
This commit adds a new method to the _FeatureDecisionService_ to check whether
the `MaatPostAssessmentProcessing` flag is enabled for the current user.

As part of this commit, the tests have been refactored to remove duplication
which would otherwise occur from new feature flag checking methods being added
to the _FeatureDecisionService_, given that the methods themselves only differ
in the feature flag and required action being checked. When adding a new method
to this class, only the parameter sets needs to be updated in the corresponding
tests.