Run request validation checks in WorkflowPreProcessorService
This commit updates the `preProcessRequest` method of the
_WorkflowPreProcessorService_ to run validation checks against
the workflow request. Specifically, it verifies the application
status and timestamp, as well as checking that the user has the
requisite permissions to perform the requested action.

This request validation is added to the hardship and means assessment
flows and is gated behind a feature flag in the means assessment flow
specifically, where the legacy stored procedure was previously called.