Rename CCP stub methods to make intention clear
This commit renames the stub methods for the Crown Court
Proceedings Service used to update proceedings, to make
it clearer what these methods are stubbing. This is
especially true for the update method available at
`api/internal/v1/proceedings/update-crown-court` within
CCP, which is not updateing the Crown Court but is actually
updating the Crown Court Outcome.