Update RepOrder date modified when updating application
This commit updates the `dateModified` date of a Rep Order whenever
any changes are made to the application it represents, be they changes
to income evidence, hardship reviews or means assessments.

This change is made so that the `dateModified` represents the accurate
point at which any changes were made to a Rep Order, allowing this date
to then be used in MAAT before a Rep Order is reserved to determine if
the user has the latest data loaded. Given that the user could be on a
number of different pages in MAAT when they attempt to reserve the Rep
Order, all possible routes to updating a Rep Order need to be covered,
which is what is done in this commit.