Fix security vulnerability in spring-security-crypto
This commit fixes the [recently published security vulnerability]
(https://security.snyk.io/vuln/SNYK-JAVA-ORGSPRINGFRAMEWORKSECURITY-9486467)
in `spring-security-crypto` by forcibly bumping the dependency up to version
`6.4.4`.

Spring Boot version `3.4.4` does patch `spring-security-crypto` to this same
version, but in order to upgrade Spring Boot, Crime Commons modules also need
to be upgraded at the same time, which has more potential to cause problems
elsewhere. In that context and to ensure that the Orchestration Service
can be as quickly fixed to allow deployments once again, this temporary fix
is made in this commit until such time as Spring Boot is upgraded to `3.4.4`.