Fix inconsistencies with Rep Order date modified
This commit fixes inconsistencies introduced regarding Rep Order
date modified timestamps and the timestamp returned as part of
the `ApplicationDTO` object. Previously, the date modified field
of the Rep Order was being updated but the `timestamp` field of the
`ApplicationDTO` was not altered, resulting in the caller receiving
the previous date modified timestamp and not the updated one.