Refactor and centralise date modified update logic
This commit refactors the date modified update logic into
a new service, wrapping the call to the _RepOrderService_
as part of this, in an effort to remove duplication.

A number of options were considered, including updating the base
class _AssessmentOrchestrator_ to make it abstract instead of an
interface and include the concrete method there, but Lombok does
support the instantiation of sub-classes with the @RequiredArgsConstructor
annotation where the base class does not have a parameterless constructor.

Due to this limitation and also thinking about composition over
inheritance, a new _AssessmentService_ has instead been created
and composed into existing services where we were previously calling
the _RepOrderService_ to update the date modified.

It should be noted that this logic has not been simply added to the
_RepOrderService_, as easy as that would be, because the _ApplicationDTO_
is strictly separate from a Rep Order and would also cause us to
violate the single responsibility principle.