Use correct timestamp set in Rep Order by Hibernate in ApplicationDTO
This commit updates the _RepOrderService_ to get and return the latest
_RepOrderDTO_ immediately after updating the date modified on the Rep
Order.

The reason for this change is because Hibernate automatically updates
the date modified field of the Rep Order entity whenever there are _any_
changes to said Rep Order; this is resulting in a situation at the moment
where a request to patch the Rep Order in the Court Data API with a new
date modified first updates the field to the value we pass, but is then
updated immediately after by Hibernate because the Rep Order entity was
changed.

This is a separate issue in its own right and a [task has been captured
to return the updated _RepOrderDTO_ from the Court Data API endpoint]
(https://dsdmoj.atlassian.net/browse/LCAM-1806) that we are calling to
patch the Rep Order. Until that work is done, the most straightward and
quickest fix to this issue (which is causing API test failures) is to
manually get the latest Rep Order after we update the date modified,
then use the timestamp from this entity when setting the application
timestamp returned upstream.