Update assessment summaries on IoJ Appeal creation
This commit adds logic to the IojAppealsOrchestrationService to
update the assessment summary object of the ApplicationDTO to
add the newly created IoJ Appeal information after.

This commit also adds a new method to the AssessmentSummaryService
to obtain an AssessmentSummaryDTO from an IojAppealDTO and updates
tests elsewhere as we now have an ambiguous method signature where
the parameter was not previously specified in the means assessment
orchestration service tests.