Only update ApplicationDTO when mags rep MagsDecisionResult exists
This commit updates the handling of the call to determine the mags
rep decision result to only update the ApplicationDTO on the
workflow request object if the decision returned by CCP is not null.
It would be null in the case where the validation against the case
type and assessment status does not match what is required to
calculate the mags rep decision, in which case to align with the
old stored procedures this should not be set on the application.