Switch CircleCI to use full checkout
This commit switches the build and test step of our
CircleCI CI pipeline to use the `full` git checkout
mode rather than `blobless` git checkout, which became
the default for pipelines on November 3rd 2025 and has
been rolled out to orgs on a gradual basis. This has
affected Orchestration as SonarQube is now failing
to run.